Add console formatting and shuffle unit test order
Shuffling is added because there may be a successful unit test in one instance,
and not another. This is because a particular unit test may leave the main
objects in an inconsistent state that other unit tests may rely on. By
randomizing the order instead of using a static order, we increase the chance
from 0% to partial chance of finding these bugs. (this happened in PUDL)